Split REST tests into smaller files
