+++ rest/tests/tests.py
@@ -1,19 +1,11 @@
 import math
 import tempfile
 import os.path
-import gzip
-import requests
-import re
-from collections import defaultdict
-from pathlib import Path
 
 from django.core.files import File as DjangoFile
-from django.core.files.uploadedfile import SimpleUploadedFile
 from django.test import override_settings
 from rest_framework import status
 from rest_framework.test import APITestCase
-
-from goatools.obo_parser import GODag
 
 from app.models import (
     Species,
@@ -30,7 +22,6 @@ from app.models import (
     MetacellLink,
     MetacellGeneExpression,
     SAMap,
-    GlobalFile,
     SpeciesFile,
 )
 
@@ -337,597 +328,6 @@ class OrthologsTests(APITestCase):
         self.assertEqual(ortholog_counts[0]["gene_count"], 4)
 
 
-class GeneModulesData(APITestCase):
-    """Data to test GeneModule-associated endpoints"""
-
-    @classmethod
-    def setUpTestData(cls):
-        cls.species1 = Species.objects.create(scientific_name="species3")
-        cls.dataset1 = cls.species1.datasets.create(name="dataset3")
-
-        # Create modules with different number of genes
-        cls.module1 = cls.dataset1.gene_modules.create(name="module_xyz")
-        cls.module2 = cls.dataset1.gene_modules.create(name="module_abc")
-        cls.module3 = cls.dataset1.gene_modules.create(name="module_123")
-        cls.module4 = cls.dataset1.gene_modules.create(name="module_000")
-
-        # Prepare transcription factor gene list
-        tfs = GeneList.objects.create(name="Transcription factors")
-
-        # Set up genes and their membership
-        scores = [0.09, 0.51, 0.14, 0.53, 0.16, 0.05, 0.05, 0.78, 0.23, 0.64, 0.12, 0.95]
-        for i, score in enumerate(scores, start=1):
-            gene = cls.species1.genes.create(name=f"gene{i}")
-
-            # First 3 genes go to module1, rest to module2
-            if i <= 3:
-                m = cls.module1
-            elif i == 4:
-                m = cls.module2
-            else:
-                m = cls.module3
-            m.membership.create(gene=gene, membership_score=score)
-
-            # Label some genes as TFs
-            if i in [3, 5, 7, 8]:
-                tfs.genes.add(gene)
-
-        cls.genes = list(cls.species1.genes.all())
-
-        # Dataset without gene modules
-        species2 = Species.objects.create(scientific_name="species4")
-        species2.datasets.create(name="dataset4")
-
-
-class GeneModulesTests(GeneModulesData):
-    """Tests GeneModules and GeneModuleMembership endpoint"""
-
-    def test_retrieve_modules(self):
-        url = "/api/v1/modules/?dataset=species3-dataset3"
-        response = self.client.get(url, format="json")
-        modules = response.data["results"]
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(modules), 4)
-
-        # Test modules with different number of genes
-        expected = [
-            # (name, gene_count, top_tf, gene_hubs)
-            ("module_000", 0, set(), set()),
-            ("module_123", 8, {"gene5", "gene7", "gene8"}, {"gene5", "gene8", "gene9", "gene10", "gene12"}),
-            ("module_abc", 1, set(), {"gene4"}),
-            ("module_xyz", 3, {"gene3"}, {"gene1", "gene2", "gene3"}),
-        ]
-
-        for m, (name, gene_count, top_tf, gene_hubs) in zip(modules, expected):
-            self.assertEqual(m["module"], name)
-            self.assertEqual(m["gene_count"], gene_count)
-            self.assertSetEqual(set(m["top_tf"]), top_tf)
-            self.assertSetEqual(set(m["gene_hubs"]), gene_hubs)
-
-    def test_retrieve_ordered_modules(self):
-        url = "/api/v1/modules/?dataset=species3-dataset3&order_by_gene_count=true"
-        response = self.client.get(url, format="json")
-        modules = response.data["results"]
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(modules), 4)
-
-        # Test sorted modules
-        expected = [
-            # (name, gene_count)
-            ("module_123", 8),
-            ("module_xyz", 3),
-            ("module_abc", 1),
-            ("module_000", 0),
-        ]
-
-        for m, (name, gene_count) in zip(modules, expected):
-            self.assertEqual(m["module"], name)
-            self.assertEqual(m["gene_count"], gene_count)
-
-    def test_retrieve_empty_module(self):
-        url = "/api/v1/modules/?dataset=species4-dataset4"
-        response = self.client.get(url, format="json")
-        modules = response.data["results"]
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(modules, [])
-
-    def test_retrieve_module_membership(self):
-        limit = 3
-        module = "module_123"
-        dataset = "species3-dataset3"
-
-        url = f"/api/v1/module_membership/?dataset={dataset}&module={module}&limit={limit}"
-        response = self.client.get(url, format="json")
-        membership = response.data["results"]
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(membership), 3)
-
-        # Test module membership
-        expected = [
-            # (name, gene, score)
-            ("module_123", "gene6", "0.050"),
-            ("module_123", "gene7", "0.050"),
-            ("module_123", "gene5", "0.160"),
-        ]
-
-        for m, (name, gene, score) in zip(membership, expected):
-            self.assertEqual(m["dataset"], dataset)
-            self.assertEqual(m["module"], name)
-            self.assertEqual(m["gene"], gene)
-            self.assertEqual(m["score"], score)
-
-
-class GeneModuleSimilarity(GeneModulesData):
-    """Tests GeneModuleSimilarity and GeneModuleSimilarityGenes endpoints"""
-
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-
-        # Add two genes from module3 to module2
-        genes = list(cls.module3.genes.all())
-        cls.module2.membership.create(gene=genes[0], membership_score=0.5)
-        cls.module2.membership.create(gene=genes[1], membership_score=0.1)
-
-        # Create different dataset within same species
-        cls.d1 = cls.dataset1
-        cls.d2 = cls.dataset1.species.datasets.create(name="dataset4")
-
-        # Assign genes to modules
-        d2_module1 = cls.d2.gene_modules.create(name="module_blue")
-        d2_module2 = cls.d2.gene_modules.create(name="module_yellow")
-        cls.d2.gene_modules.create(name="module_orange")
-        d2_module4 = cls.d2.gene_modules.create(name="module_green")
-        cls.d2_module1 = d2_module1
-
-        genes = cls.dataset1.species.genes.all()
-        [d2_module1.membership.create(gene=genes[i]) for i in [1, 4, 5, 7, 8]]
-        [d2_module2.membership.create(gene=genes[i]) for i in [3, 5, 9, 10, 11]]
-        # d2_module3: module without genes
-        [d2_module4.membership.create(gene=genes[i]) for i in [2, 3, 4, 5, 6, 7, 9]]
-
-        # Create a dataset in a different species
-        s3 = Species.objects.create(scientific_name="novel")
-        cls.d3 = s3.datasets.create(name="polyp")
-
-        [s3.genes.create(name=f"gene{i}") for i in list("ABCDEFGHIJ")]
-        s3_genes = s3.genes.all()
-
-        [cls.d3.gene_modules.create(name=f"mod{i}") for i in ["alpha", "beta", "gamma"]]
-
-        [cls.d3.gene_modules.first().membership.create(gene=s3_genes[i]) for i in [1, 4, 5, 7, 8]]
-        # d3_module2: module without genes
-        [cls.d3.gene_modules.last().membership.create(gene=s3_genes[i]) for i in [2, 3, 7, 8, 9]]
-
-        # Create orthologs to intersect genes for the 'modgamma' module
-        og1 = Orthogroup.objects.create(name="og1")
-        og2 = Orthogroup.objects.create(name="og2")
-        og3 = Orthogroup.objects.create(name="og3")
-        og4 = Orthogroup.objects.create(name="og4")
-
-        genes[3].orthologs.create(species=cls.d2.species, orthogroup=og1)
-        genes[4].orthologs.create(species=cls.d2.species, orthogroup=og3)
-        s3_genes[5].orthologs.create(species=s3, orthogroup=og1)
-        s3_genes[6].orthologs.create(species=s3, orthogroup=og2)
-        s3_genes[7].orthologs.create(species=s3, orthogroup=og3)
-
-        # Create gene with multiple orthogroups to test similarity comparison case
-        s3_genes[8].orthologs.create(species=s3, orthogroup=og3)
-        s3_genes[8].orthologs.create(species=s3, orthogroup=og4)
-
-    def run_similarity_tests(self, sim, expected):
-        for m, (module, module2, uniq, uniq2, shared, *shared_split) in zip(sim, expected):
-            if shared_split:
-                # Jaccard similarity index via othorgroups
-                shared1, shared2 = shared_split
-            else:
-                # Jaccard similarity index via genes directly
-                shared1 = shared
-                shared2 = shared
-
-            # Calculate Jaccard similarity index in percentage
-            total = uniq + uniq2 + shared
-            jaccard = round(shared / total, 2) if total else 0
-
-            self.assertEqual(m["module"], module)
-            self.assertEqual(m["module2"], module2)
-            self.assertEqual(m["similarity"], jaccard)
-            self.assertEqual(m["unique_genes_module"], uniq)
-            self.assertEqual(m["unique_genes_module2"], uniq2)
-            self.assertEqual(m["shared_genes_module"], shared1)
-            self.assertEqual(m["shared_genes_module2"], shared2)
-
-    def run_similarity_genes_tests(self, sim, expected):
-        for m, (type, dataset, module, gene, og) in zip(sim, expected):
-            self.assertEqual(m["overlap"], type)
-            self.assertEqual(m["dataset"], dataset)
-            self.assertEqual(m["module"], module)
-            self.assertEqual(m["gene"], gene)
-            self.assertEqual(m["orthogroups"], og)
-
-    def group_genes(self, sim):
-        grouped = defaultdict(set)
-        for row in sim:
-            grouped[row["overlap"]].add(row["gene"])
-        return dict(grouped)
-
-    def test_retrieve(self):
-        dataset = "species3-dataset3"
-
-        url = f"/api/v1/module_similarity/?dataset={dataset}"
-        response = self.client.get(url)
-        sim = response.data
-
-        # Test module eigengenes
-        expected = [
-            # (module, module2, unique_genes_module, unique_genes_module2, shared)
-            ("module_000", "module_123", 0, 8, 0),
-            ("module_000", "module_abc", 0, 3, 0),
-            ("module_000", "module_xyz", 0, 3, 0),
-            ("module_123", "module_abc", 6, 1, 2),
-            ("module_123", "module_xyz", 8, 3, 0),
-            ("module_abc", "module_xyz", 3, 3, 0),
-        ]
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(sim), len(expected))
-        self.run_similarity_tests(sim, expected)
-
-    def test_nonexisting_module(self):
-        """Raise error when using non-existing modules."""
-        dataset = "species3-dataset3"
-        module = "module_123"
-
-        # Raise error for a non-existing module
-        module2 = "not_a_real_module"
-        url = f"/api/v1/module_similarity/?dataset={dataset}&module={module}&module2={module2}"
-        with self.assertRaises(ValueError) as context:
-            self.client.get(url)
-        self.assertIn("module not_a_real_module does not exist", str(context.exception))
-
-    def test_filter(self):
-        """Filter gene modules."""
-        dataset = "species3-dataset3"
-        module = "module_123"
-        module2 = "module_abc"
-
-        url = f"/api/v1/module_similarity/?dataset={dataset}&module={module}&module2={module2}"
-        response = self.client.get(url)
-        sim = response.data
-
-        expected = [
-            # (module, module2, unique_genes_module, unique_genes_module2, shared)
-            ("module_123", "module_abc", 6, 1, 2),
-        ]
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(sim), len(expected))
-        self.run_similarity_tests(sim, expected)
-
-    def test_compare_within_species_genes(self):
-        """Retrieve list of genes."""
-        dataset = "species3-dataset3"
-        module = "module_123"
-        module2 = "module_abc"
-
-        url = f"/api/v1/module_similarity_genes/?dataset={dataset}&module={module}&module2={module2}"
-        response = self.client.get(url)
-        sim = response.data
-
-        # Test module eigengenes
-        expected = [
-            # (overlap, dataset, module, gene, orthogroup)
-            (f"unique_{dataset}_{module}", dataset, module, "gene5", []),
-            (f"unique_{dataset}_{module}", dataset, module, "gene6", []),
-            (f"unique_{dataset}_{module}", dataset, module, "gene7", []),
-            (f"unique_{dataset}_{module}", dataset, module, "gene8", []),
-            (f"unique_{dataset}_{module}", dataset, module, "gene9", []),
-            (f"unique_{dataset}_{module}", dataset, module, "gene12", ["og1"]),
-            (f"unique_{dataset}_{module2}", dataset, module2, "gene4", []),
-            ("shared", None, None, "gene10", []),
-            ("shared", None, None, "gene11", []),
-        ]
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(sim), len(expected))
-        self.run_similarity_genes_tests(sim, expected)
-
-        genes_123 = {"gene5", "gene6", "gene7", "gene8", "gene9", "gene12"}
-        genes_abc = {"gene4"}
-        genes_both = {"gene10", "gene11"}
-
-        result = self.group_genes(sim)
-        self.assertEqual(len(result), 3)
-        self.assertSetEqual(result["unique_species3-dataset3_module_123"], genes_123)
-        self.assertSetEqual(result["unique_species3-dataset3_module_abc"], genes_abc)
-        self.assertSetEqual(result["shared"], genes_both)
-
-    def test_compare_within_species_genes_no_modules(self):
-        """Raise error when listing genes without module parameters."""
-        dataset = "species3-dataset3"
-
-        # Raise error for not inputting modules to compare
-        url = f"/api/v1/module_similarity_genes/?dataset={dataset}"
-        with self.assertRaises(ValueError) as context:
-            self.client.get(url)
-        self.assertIn("please define 'module' and 'module2' parameters", str(context.exception))
-
-    def test_compare_same_species(self):
-        """Compare datasets from same species."""
-        dataset = self.d1.slug
-        dataset2 = self.d2.slug
-        self.assertEqual(self.d1.species, self.d2.species)
-
-        url = f"/api/v1/module_similarity/?dataset={dataset}&dataset2={dataset2}"
-        response = self.client.get(url)
-        sim = response.data
-
-        expected = [
-            # (module, module2, unique_genes_module, unique_genes_module2, shared)
-            ("module_000", "module_blue", 0, 5, 0),
-            ("module_000", "module_green", 0, 7, 0),
-            ("module_000", "module_orange", 0, 0, 0),
-            ("module_000", "module_yellow", 0, 5, 0),
-            ("module_123", "module_blue", 5, 2, 3),
-            ("module_123", "module_green", 4, 3, 4),
-            ("module_123", "module_orange", 8, 0, 0),
-            ("module_123", "module_yellow", 4, 1, 4),
-            ("module_abc", "module_blue", 2, 4, 1),
-            ("module_abc", "module_green", 1, 5, 2),
-            ("module_abc", "module_orange", 3, 0, 0),
-            ("module_abc", "module_yellow", 3, 5, 0),
-            ("module_xyz", "module_blue", 1, 3, 2),
-            ("module_xyz", "module_green", 1, 5, 2),
-            ("module_xyz", "module_orange", 3, 0, 0),
-            ("module_xyz", "module_yellow", 2, 4, 1),
-        ]
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(sim), len(expected))
-        self.run_similarity_tests(sim, expected)
-
-    def test_compare_same_species_sorted(self):
-        """Compare datasets from same species."""
-        sort_modules = "true"
-        dataset = self.d1.slug
-        dataset2 = self.d2.slug
-        self.assertEqual(self.d1.species, self.d2.species)
-
-        url = f"/api/v1/module_similarity/?dataset={dataset}&dataset2={dataset2}&sort_modules={sort_modules}"
-        response = self.client.get(url)
-        sim = response.data
-
-        expected = [
-            # (module, module2, unique_genes_module, unique_genes_module2, shared)
-            ("module_123", "module_yellow", 4, 1, 4),
-            ("module_123", "module_green", 4, 3, 4),
-            ("module_xyz", "module_blue", 1, 3, 2),
-            ("module_123", "module_blue", 5, 2, 3),
-            ("module_abc", "module_green", 1, 5, 2),
-            ("module_xyz", "module_green", 1, 5, 2),
-            ("module_abc", "module_blue", 2, 4, 1),
-            ("module_xyz", "module_yellow", 2, 4, 1),
-            ("module_000", "module_blue", 0, 5, 0),
-            ("module_000", "module_green", 0, 7, 0),
-            ("module_000", "module_orange", 0, 0, 0),
-            ("module_000", "module_yellow", 0, 5, 0),
-            ("module_123", "module_orange", 8, 0, 0),
-            ("module_abc", "module_orange", 3, 0, 0),
-            ("module_abc", "module_yellow", 3, 5, 0),
-            ("module_xyz", "module_orange", 3, 0, 0),
-        ]
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(sim), len(expected))
-        self.run_similarity_tests(sim, expected)
-
-    def test_compare_same_species_genes(self):
-        """Compare genes from same species with module filtering."""
-        dataset = self.d1.slug
-        dataset2 = self.d2.slug
-        self.assertEqual(self.d1.species, self.d2.species)
-
-        # Check genes for this module
-        ma = self.module3
-        self.assertEqual(ma.name, "module_123")
-        genes1 = {"gene5", "gene6", "gene7", "gene8", "gene9", "gene10", "gene11", "gene12"}
-        self.assertSetEqual(set(ma.genes.values_list("name", flat=True)), genes1)
-
-        # Check genes for this module
-        mb = self.d2_module1
-        self.assertEqual(mb.name, "module_blue")
-        genes2 = {"gene2", "gene3", "gene5", "gene6", "gene10"}
-        self.assertSetEqual(set(mb.genes.values_list("name", flat=True)), genes2)
-
-        url = (
-            f"/api/v1/module_similarity_genes/?dataset={dataset}&dataset2={dataset2}&module={ma.name}&module2={mb.name}"
-        )
-        response = self.client.get(url)
-        sim = response.data
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        # Test results
-        self.assertEqual(len(sim), len(genes1 | genes2))
-
-        result = self.group_genes(sim)
-        self.assertEqual(len(result), 3)
-        self.assertSetEqual(result[f"unique_{dataset}_{ma.name}"], genes1 - genes2)
-        self.assertSetEqual(result[f"unique_{dataset2}_{mb.name}"], genes2 - genes1)
-        self.assertSetEqual(result["shared"], genes1 & genes2)
-
-    def test_compare_different_species(self):
-        """Compare datasets from different species via orthologs."""
-        dataset = self.d2.slug
-        dataset2 = self.d3.slug
-        self.assertNotEqual(self.d2.species, self.d3.species)
-
-        url = f"/api/v1/module_similarity/?dataset={dataset}&dataset2={dataset2}"
-        response = self.client.get(url)
-        sim = response.data
-
-        expected = [
-            # (module, module2, unique_genes_module, unique_genes_module2, shared, shared1, shared2)
-            ("module_blue", "modalpha", 4, 3, 3, 1, 2),
-            ("module_blue", "modbeta", 5, 0, 0, 0, 0),
-            ("module_blue", "modgamma", 4, 3, 3, 1, 2),
-            ("module_green", "modalpha", 5, 2, 5, 2, 3),
-            ("module_green", "modbeta", 7, 0, 0, 0, 0),
-            ("module_green", "modgamma", 6, 3, 3, 1, 2),
-            ("module_orange", "modalpha", 0, 5, 0, 0, 0),
-            ("module_orange", "modbeta", 0, 0, 0, 0, 0),
-            ("module_orange", "modgamma", 0, 5, 0, 0, 0),
-            ("module_yellow", "modalpha", 4, 4, 2, 1, 1),
-            ("module_yellow", "modbeta", 5, 0, 0, 0, 0),
-            ("module_yellow", "modgamma", 5, 5, 0, 0, 0),
-        ]
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(sim), len(expected))
-        self.run_similarity_tests(sim, expected)
-
-    def test_compare_different_species_genes(self):
-        """Compare datasets from different species via orthologs."""
-        dataset = self.d2.slug
-        dataset2 = self.d3.slug
-        self.assertNotEqual(self.d2.species, self.d3.species)
-
-        # Check genes for this module
-        ma = self.d2_module1
-        self.assertEqual(ma.name, "module_blue")
-        genes1 = {"gene2", "gene3", "gene5", "gene6", "gene10"}
-        self.assertSetEqual(set(ma.genes.values_list("name", flat=True)), genes1)
-
-        # Check genes for this module
-        mb = self.d3.gene_modules.get(name="modgamma")
-        self.assertEqual(mb.name, "modgamma")
-        genes2 = {"geneC", "geneD", "geneH", "geneI", "geneJ"}
-        self.assertSetEqual(set(mb.genes.values_list("name", flat=True)), genes2)
-
-        url = (
-            f"/api/v1/module_similarity_genes/?dataset={dataset}&dataset2={dataset2}&module={ma.name}&module2={mb.name}"
-        )
-        response = self.client.get(url)
-        sim = response.data
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        # Check orthologues
-        genes1_ogs = set(ma.genes.filter(orthogroups__isnull=False).values_list("name", flat=True))
-        self.assertSetEqual(genes1_ogs, {"gene2"})
-        genes2_ogs = set(mb.genes.filter(orthogroups__isnull=False).values_list("name", flat=True))
-        self.assertSetEqual(genes2_ogs, {"geneH", "geneI"})
-
-        # Test results
-        self.assertEqual(len(sim), len(genes1 | genes2))
-
-        result = self.group_genes(sim)
-        self.assertEqual(len(result), 4)
-        self.assertSetEqual(result[f"unique_{dataset}_{ma.name}"], genes1 - genes1_ogs)
-        self.assertSetEqual(result[f"unique_{dataset2}_{mb.name}"], genes2 - genes2_ogs)
-        self.assertSetEqual(result[f"shared_{dataset}_{ma.name}"], genes1_ogs)
-        self.assertSetEqual(result[f"shared_{dataset2}_{mb.name}"], genes2_ogs)
-
-
-class GeneModuleEigengene(GeneModulesData):
-    """Tests GeneModuleEigengene endpoint"""
-
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-        m1 = cls.dataset1.metacells.create(name="1", x=0.2, y=0.5)
-        m2 = cls.dataset1.metacells.create(name="2", x=0.6, y=0.2)
-        m3 = cls.dataset1.metacells.create(name="3", x=0.1, y=0.8)
-
-        cls.module1.eigengene_values.create(metacell=m1, eigengene_value=0.167)
-        cls.module1.eigengene_values.create(metacell=m2, eigengene_value=0.135)
-        cls.module1.eigengene_values.create(metacell=m3, eigengene_value=0.153)
-
-        cls.module2.eigengene_values.create(metacell=m1, eigengene_value=0.696)
-        cls.module2.eigengene_values.create(metacell=m2, eigengene_value=0.897)
-        cls.module2.eigengene_values.create(metacell=m3, eigengene_value=0.336)
-
-    def test_retrieve_module_eigengenes(self):
-        dataset = "species3-dataset3"
-
-        url = f"/api/v1/module_eigengenes/?dataset={dataset}"
-        response = self.client.get(url)
-        eigengenes = response.data["results"]
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(eigengenes), 6)
-
-        # Test module eigengenes
-        expected = [
-            # (name, metacell, eigengene_values)
-            ("module_xyz", "1", "0.167"),
-            ("module_xyz", "2", "0.135"),
-            ("module_xyz", "3", "0.153"),
-            ("module_abc", "1", "0.696"),
-            ("module_abc", "2", "0.897"),
-            ("module_abc", "3", "0.336"),
-        ]
-
-        for m, (module, metacell, value) in zip(eigengenes, expected):
-            self.assertEqual(m["dataset"], dataset)
-            self.assertEqual(m["module"], module)
-            self.assertEqual(m["metacell_name"], metacell)
-            self.assertEqual(m["eigengene_value"], value)
-
-    def test_retrieve_single_module_eigengene(self):
-        module = "module_xyz"
-        dataset = "species3-dataset3"
-
-        url = f"/api/v1/module_eigengenes/?dataset={dataset}&module={module}"
-        response = self.client.get(url)
-        eigengene_values = response.data["results"]
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(eigengene_values), 3)
-
-        # Test module eigengenes
-        expected = [
-            # (name, metacell, value)
-            (module, "1", "0.167"),
-            (module, "2", "0.135"),
-            (module, "3", "0.153"),
-        ]
-
-        for m, (module, metacell, value) in zip(eigengene_values, expected):
-            self.assertEqual(m["dataset"], dataset)
-            self.assertEqual(m["module"], module)
-            self.assertEqual(m["metacell_name"], metacell)
-            self.assertEqual(m["eigengene_value"], value)
-
-    def test_retrieve_sorted_module_eigengenes(self):
-        module = "module_xyz"
-        dataset = "species3-dataset3"
-        sort_modules = "true"
-
-        url = f"/api/v1/module_eigengenes/?dataset={dataset}&sort_modules={sort_modules}"
-        response = self.client.get(url)
-        eigengenes = response.data["results"]
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(eigengenes), 6)
-
-        # Test module eigengenes
-        expected = [
-            # (name, metacell, value)
-            ("module_abc", "1", "0.696"),
-            ("module_abc", "2", "0.897"),
-            ("module_abc", "3", "0.336"),
-            ("module_xyz", "1", "0.167"),
-            ("module_xyz", "2", "0.135"),
-            ("module_xyz", "3", "0.153"),
-        ]
-
-        for m, (module, metacell, value) in zip(eigengenes, expected):
-            self.assertEqual(m["dataset"], dataset)
-            self.assertEqual(m["module"], module)
-            self.assertEqual(m["metacell_name"], metacell)
-            self.assertEqual(m["eigengene_value"], value)
-
-
 class SAMapTests(APITestCase):
     """Tests SAMap endpoint"""
 
@@ -1000,220 +400,3 @@ class AlignTests(APITestCase):
         response = self.client.post(url, data, format="json")
 
         self.check_expected_alignment(response)
-
-
-@override_settings(MEDIA_ROOT=tempfile.mkdtemp())
-class EnrichmentAnalysisTests(APITestCase):
-    """Tests enrichment endpoint."""
-
-    def extract_go_from_emapper_file(
-        self, obo_path, file_path=Path(__file__).parent / "test_fixtures" / "emapper_annotation.txt.gz"
-    ):
-        """Extract all GO terms from annotation file."""
-
-        # Expand GO terms from OBO file to include all parent terms
-        go_dag = GODag(obo_path, load_obsolete=True)
-
-        go_ids = set()
-        pattern = re.compile(r"GO:\d{7}")
-
-        with gzip.open(file_path, "rt") as f:
-            for line in f:
-                for go_id in pattern.findall(line):
-                    go_ids.add(go_id)
-                    go_ids.update(go_dag[go_id].get_all_parents())
-
-        return go_ids
-
-    def download_go_obo(self, output_path=Path(__file__).parent / "test_fixtures" / "go-basic.obo"):
-        """Download GO basic OBO file."""
-        if output_path.exists():
-            return output_path
-
-        url = "https://purl.obolibrary.org/obo/go/go-basic.obo"
-        with requests.get(url, stream=True) as r:
-            r.raise_for_status()
-            with open(output_path, "wb") as f:
-                for chunk in r.iter_content(chunk_size=1024 * 1024):  # 1MB chunks
-                    if chunk:
-                        f.write(chunk)
-
-        return output_path
-
-    def prepare_filtered_go_obo(self, output_path=Path(__file__).parent / "test_fixtures" / "go-basic-subset.obo"):
-        """Prepare a subset GO OBO file for testing based on a list of GO terms to keep."""
-
-        obo_path = self.download_go_obo()
-        go_ids = self.extract_go_from_emapper_file(obo_path)
-
-        # Ignore unneeded lines with the following prefixes to reduce file size
-        skip_line_prefixes = ("def", "comment", "xref", "synonym", "property_value", "consider", "subset", "alt_id")
-
-        gzip_output_path = str(output_path) + ".gz"
-        with open(obo_path, "r") as r, gzip.open(gzip_output_path, "wt") as out:
-            block = []
-            go_block = False
-            current_id = None
-
-            for line in r:
-                line = line.strip()
-                if not line:
-                    continue
-
-                if line.split(":", 1)[0] in skip_line_prefixes:
-                    continue
-
-                if line.startswith("[Term]"):
-                    if block and go_block:
-                        out.write("\n".join(block) + "\n\n")
-
-                    block = [line]
-                    go_block = False
-                    current_id = None
-                    continue
-
-                block.append(line)
-                if line.startswith("id:"):
-                    current_id = line.split()[1]
-                    if current_id in go_ids:
-                        go_block = True
-
-            # last block
-            if block and go_block:
-                out.write("\n".join(block) + "\n")
-
-        # Remove original OBO file
-        obo_path.unlink()
-
-        return gzip_output_path
-
-    @classmethod
-    def setUpTestData(cls):
-        aque = Species.objects.create(scientific_name="Amphimedon queenslandica")
-        aque_adult = aque.datasets.create(name="adult")
-
-        # Get OBO file with GO term definitions
-        go_obo_compressed = Path(__file__).parent / "test_fixtures" / "go-basic-subset.obo.gz"
-        with gzip.open(go_obo_compressed, "rb") as f:
-            go_obo_file = f.read()
-        cls.go_obo = GlobalFile.objects.create(
-            type="go-basic-obo", file=SimpleUploadedFile("go-basic.obo", go_obo_file)
-        )
-
-        # Get functional annotation
-        emapper_file = Path(__file__).parent / "test_fixtures" / "emapper_annotation.txt.gz"
-        cls.emapper = aque.files.create(
-            type="eggnog-mapper", file=SimpleUploadedFile("emapper.txt.gz", emapper_file.read_bytes())
-        )
-
-        # Prepare background genes: add genes to metacell gene expression based on functional annotation
-        mc1 = aque_adult.metacells.create(name="mc1", x="1", y="5")
-        cls.genes = []
-        with gzip.open(emapper_file, "rt") as f:
-            for line in f:
-                if line.startswith("#") or not line.strip():
-                    continue
-                gene = line.split("\t")[0]
-                g = aque.genes.create(name=gene)
-                aque_adult.mge.create(gene=g, metacell=mc1)
-                cls.genes.append(gene)
-
-    def check_enrichment_response(self, response, genes, obsolete=False):
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        for d in response.data:
-            self.assertRegex(d["term"], r"^GO:\d{7}$", "Expected GO term name")
-            self.assertIn(d["enrichment"], {"e", "p"}, "Enrichment results")
-
-            if obsolete:
-                self.assertGreaterEqual(d["depth"], 0, "GO term depth ≥ 0")
-
-                # Check if returning any obsolete GO term
-                self.assertTrue(
-                    any(d.get("is_obsolete") for d in response.data), "Expected at least one obsolete term in response"
-                )
-
-                # Check if the obsolete status matches the OBO definition
-                go_obo = GODag(self.go_obo.file.path, load_obsolete=True, prt=None)
-                for d in response.data:
-                    self.assertEqual(go_obo[d["term"]].is_obsolete, d.get("is_obsolete"))
-            else:
-                self.assertGreaterEqual(d["depth"], 1, "GO term depth ≥ 1")
-                self.assertNotIn("is_obsolete", d, "No obsolete field")
-
-            self.assertSetEqual(set(d["genes"]), genes, "Genes match input in these examples")
-            self.assertEqual(d["query_count"], len(d["genes"]), "Match number of genes")
-            self.assertEqual(d["query_hit_count"], len(d["genes"]), "Match number of genes")
-            self.assertEqual(d["background_hit_count"], len(d["genes"]), "Match number of genes")
-            self.assertEqual(d["background_count"], len(self.genes), "Match all genes")
-
-            self.assertTrue(0 <= d["pvalue"] <= 1, "Expected p-value")
-            self.assertTrue(0 <= d["qvalue"] <= 1, "Expected adjusted p-value")
-            self.assertLess(d["pvalue"], 0.05, "Significant p-value")
-            self.assertLess(d["qvalue"], 0.05, "Significant q-value")
-            self.assertLessEqual(d["pvalue"], d["qvalue"], "p-value ≤ adjusted p-value")
-
-            self.assertIn(d["namespace"], {"BP", "CC", "MF"}, "Match specific namespaces")
-            self.assertTrue(
-                all(isinstance(x, float) for x in d["similarity_coords"]),
-                "similarity_coords is a tuple of two numbers",
-            )
-
-    def test_post(self):
-        url = "/api/v1/enrichment/"
-        genes = {"Aque_Aqu2.1.30266_001", "Aque_Aqu2.1.30264_001", "Aque_Aqu2.1.30269_001"}
-        data = dict(dataset="amphimedon-queenslandica-adult", genes=genes)
-        response = self.client.post(url, data, format="json")
-        self.check_enrichment_response(response, genes)
-
-        for d in response.data:
-            self.assertIn(
-                d["term"],
-                {
-                    "GO:0006436",
-                    "GO:0006518",
-                    "GO:0006412",
-                    "GO:0048813",
-                    "GO:0004830",
-                    "GO:0010835",
-                    "GO:0016874",
-                    "GO:0031334",
-                    "GO:0016875",
-                    "GO:0043043",
-                    "GO:0043038",
-                    "GO:0140101",
-                    "GO:0043039",
-                },
-                "Expected GO terms",
-            )
-
-    def test_post_obsolete(self):
-        """Test if obsolete terms are included."""
-        url = "/api/v1/enrichment/"
-        genes = {"Aque_Aqu2.1.30266_001", "Aque_Aqu2.1.30264_001", "Aque_Aqu2.1.30269_001"}
-        data = dict(dataset="amphimedon-queenslandica-adult", genes=genes, obsolete=True)
-        response = self.client.post(url, data, format="json")
-        self.check_enrichment_response(response, genes, obsolete=True)
-
-        for d in response.data:
-            self.assertIn(
-                d["term"],
-                {
-                    "GO:0048813",
-                    "GO:0004830",
-                    "GO:0031334",
-                    "GO:0006412",
-                    "GO:0010835",
-                    "GO:0043604",
-                    "GO:0043603",
-                    "GO:0006436",
-                    "GO:0016875",
-                    "GO:0043038",
-                    "GO:0016874",
-                    "GO:0043043",
-                    "GO:0043039",
-                    "GO:0006518",
-                    "GO:0140101",
-                },
-                "Expected GO terms",
-            )
